feat:PersonList.get() now supports uuid argument and has been added to CallDetailRecords
To parse CDRs, the User UUID is occasionally the only identifier for the user. There was previously no easy way to look up a Person by UUID. Since this identifier is common and available in other classes, it was added to the PersonList class to allow it to be used in other places.