refactor:Recording attributes read from .details before making API call
Most of the properties in the Recording class used a .get() to see if the value existed already and, if not, made the API call to get the data. I forgot that the default in a .get() is executed regardless of whether the key was found or not, so it was wasting API calls to get data that was already present.

Rather than refactor as an if/else, I found that using a lambda that does the function call doesn't execute the pre-fetch and was easier to read.