feat:WebexLicenseList and WebexLicense via Org.licenses.
As more and more customers have multiple subscriptions and we introduce new licenses, it becomes necessary to keep track of licenses in a better way. The new classes will give a framework to handle this going forward.

Part of this effort also involved ensuring that any of the .create() and .assign_wxc() methods were able to handle the new classes rather than the dict that the property used to use. This was a big undertaking and required changes across both the Person and Workspaces (and their list classes). All the methods in the WebexLicenseList are designed to make license selection both easy and flexible.