rel:4.5.0 will move meraki dependency to optional
After installing wxcadm on a new machine with UV, I noticed that aiohttp was being installed, along with a bunch of other packages I didn't understand. The uv tree showed me that these were all dependencies of the meraki module, which almost no one uses.

I thought about going back to when I required it to be installed manually, but decided it would be good to include it as an optional dependency with wxcadm to make it easy to install with one line.