feat:Device.calling_id added and DeviceList refactor
Since the release of the PhoneOS line, devices have two IDs, one for Webex and one for Webex Calling. The API development team is planning to add a "callingId" to /v1/devices to make it easier to find the Calling ID.

This change builds the framework to support that new field once it is available and also attempts to find the "calling_id" from the /v1/telephony endpoints. Everything that uses the calling_id (i.e. DeviceLayout) has been updated to use the new, accurate value.