bug:DeviceMemberList.add() now works with ATA
Found that the API requires a t38FaxCompression bool when adding or updating members on an ATA. The code checks to see if this is a 'Cisco 191' or 'Cisco 192' and adds that. I probably need to figure out which device metadata it uses to determine that and use that instead, but this will fix the bug.